wip: ingredients are deleting but need to work on re-printing fresh view
